refactor(deauth): refactor and add test cases all deauth methods
Refactor deauth module for a better algorithm as well
as space. The following bugs where discovered during that
process:

1. The previous get_packet method would keep a local copy
of the previous packets sent. So on the second execuation of
the method it would return [previous_packets + new_packets] which
would increase exponentially.

2. On some occasions the channel for the packet could be malformed
and return garbage values. This would result in an improper deauth
packet.

Also test cases for all the methods are added so future refactoring
should be much easier.